requirements.txt
# Core Web Scraping
scrapy>=2.11.0
playwright>=1.40.0
selenium>=4.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Processing & Export
pandas>=2.1.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
Pillow>=10.0.0
reportlab>=4.0.0
matplotlib>=3.7.0
plotly>=5.17.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Automation & Scheduling
schedule>=1.2.0
APScheduler>=3.10.0

# Social Media APIs
twython>=3.9.1
praw>=7.7.0
instaloader>=4.10.0

# Database Support
pymongo>=4.5.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.1.0
SQLAlchemy>=2.0.0

# Communication & Notifications
python-telegram-bot>=20.6
# smtplib is built into Python, no need to install

# Stealth & Proxy
fake-useragent>=1.4.0
python-socks>=2.3.0
requests[socks]>=2.31.0

# Cloud Storage
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
gspread>=5.12.0
dropbox>=11.36.0

# Captcha Solving
2captcha-python>=1.1.3

# CLI & Utilities
click>=8.1.0
rich>=13.6.0
tqdm>=4.66.0

# AI and Analytics (Phase 5)
# Google Gemini API requests handled by requests library
plotly>=5.17.0
matplotlib>=3.7.0

# Web Dashboard (Phase 6)
flask>=2.3.0
gunicorn>=21.2.0

# Enhanced Libraries
fake-useragent>=1.4.0

# Development (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# GUI (optional)
customtkinter>=5.2.0

# Android Support (optional)
kivy>=2.2.0
buildozer>=1.5.0
